join reducers only into only one file and update store to not combine reducers
